Move unit test for module state to pytest
